requirements.txt
# Core MCP and Networking
mcp>=1.0.0
websockets>=11.0.3
asyncio>=3.4.3

# AI Frameworks
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
crewai>=0.1.0
crewai-tools>=0.1.0

# System and File Management
psutil>=5.9.0
pathlib>=1.0.1
watchdog>=3.0.0

# Security and Authentication
cryptography>=41.0.0
pyjwt>=2.8.0
python-decouple>=3.8

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
pydantic>=2.5.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
